organize kanban redux state/structure
