+++ src/actions/index.js
@@ -1,3 +1,5 @@
-export * from './userAuthActions';
-export * from './topicsActions';
+// export * from './userAuthActions';
+// export * from './topicsActions';
+export * from './posts';
+export * from './kanban';
